requirements.txt
# requirements.txt
streamlit>=1.27.0
openai>=1.3.0
langchain>=0.0.267
langchain-openai>=0.0.1
python-dotenv>=1.0.0
PyPDF2>=3.0.0
tiktoken>=0.5.0
pytest>=7.4.0
flake8>=6.1.0
black>=23.7.0
isort>=5.12.0
pydantic>=2.3.0